pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "entdb"
version = "1.0.0"
description = "Tenant-sharded graph database with nodes and edges"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "EntDB Team", email = "team@entdb.io"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = ["database", "graph", "tenant", "sharding", "event-sourcing"]

dependencies = []

[project.optional-dependencies]
server = [
    "aiokafka>=0.10.0",
    "aiobotocore>=2.11.0",
    "aiohttp>=3.9.0",
    "grpcio>=1.60.0",
]
sdk = [
    "aiohttp>=3.9.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
]
all = [
    "entdb[server,sdk,dev]",
]

[project.scripts]
entdb = "dbaas.entdb_server.main:main"
entdb-restore = "dbaas.entdb_server.tools.restore:main"
entdb-schema = "dbaas.entdb_server.tools.schema_cli:main"

[project.urls]
Homepage = "https://github.com/example/entdb"
Documentation = "https://github.com/example/entdb#readme"
Repository = "https://github.com/example/entdb"
Issues = "https://github.com/example/entdb/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["dbaas*", "sdk*"]

[tool.setuptools.package-data]
"*" = ["py.typed", "*.proto"]

# =============================================================================
# Ruff (linter)
# =============================================================================
[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C901",   # too complex
    "ARG001", # unused function argument
    "ARG002", # unused method argument
]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.isort]
known-first-party = ["dbaas", "sdk", "entdb_sdk", "entdb_server"]

# =============================================================================
# Mypy (type checker)
# =============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
show_error_codes = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "aiokafka.*",
    "aiobotocore.*",
    "grpc.*",
    "aiohttp.*",
]
ignore_missing_imports = true

# =============================================================================
# Pytest
# =============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require SQLite)",
    "e2e: End-to-end tests (require full stack)",
    "slow: Slow tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
timeout = 60

# =============================================================================
# Coverage
# =============================================================================
[tool.coverage.run]
source = ["dbaas", "sdk"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
show_missing = true
precision = 2
